+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 - added a wrapper for [linkit](https://github.com/dreipol/linkit) with djangocms-blog support, located at `backend.plugins.link`
 - added backend.site_config example
+- added ability to add html links and iframe to CKEditor, because there's no other way to do that in django admin
+- fixed spellchecker in ckeditor
 
 ### Documentation